points' max/min distance

+++ simple_algorithm/include/random_primality_test.h
@@ -2,6 +2,6 @@
 #define RANDOM_PRIMALITY_TEST_H
 
 // 单次调用，若输出0，则必是对的；若输出1，正确率 > 75%
-int primalityTest(long n, unsigned testTimes);
+int primalityTest(unsigned n, unsigned testTimes);
 
 #endif //RANDOM_PRIMALITY_TEST_H
